Tweak to code that sends role updates to the switches.
Send role request to switches when controller role changes if we
don't know yet if switch support the role request messages. This handles
the case where the switch connected to the controller before the
controller role had been set/enabled and thus we never sent out the
initial role request probe to the switch during the initial handshake.