Rewrite switch state management and role handling BSC-3482
This is a first big chunk of changes and work in progress. It's not
fully tested yet.
* The OFChannelHandler has be factored out into a separate file.
* State handling is now structured by the current channel state. We
implement the state machine as enum. The enum values are the discrete
states and the "events" of the state machines are member methods of the
enum values.
* Changing roles has been localized in the OFChannelHandler, we perform
timeout checking (no reply to role request) opportunisticly when we
receive any OF message. This way we don't need to complex RoleChanger
thread with task queue.