Port names should be case-insentive
Comaprisions of port names and retrivials from IOFSwitch.getPort(String)
are now case-insentivie. However, we still store port names with their
original case.

This is a quick-fix were we lower-case the port name before adding it to
the port map in PortManager. Should really use a map with
case-insensitive keys...