Makes Floodlight's Static Flow Pusher compliant with RFC 2616
Floodlight's Static Flow Pusher REST interface expects "posted data"
for DELETE requests. It uses this posted data to identify which entry
should be deleted. However, the RFC 2616 states that resources
to be deleted using the DELETE method should be identified by a
URI (not by posted data):
http://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html#sec9.7

This is a problem when you try to use a spec-compliant framework to
communicate with the REST interface provided by StaticFlowEntryPusher
module. The Restlet framework, that is being used in the implementation
of the module itself, is an example. It prevents you from posting data with
a DELETE method, according to what is specified in the HTTP1.1 RFC.