Adding initial support for automatic detection and support of different types of OpenFlow pipelines.
Currently the only switches that are supported are OVS and certain types of HP switches. More work needs to be done in order to fully support specific switches and test that flow mods are correctly being conformed to these different OpenFlow pipelines. This can be difficult to do and may require trial and error – namely due to the lack of clarity regarding OpenFlow support in many switch specifications.